requirements.txt
colorlog
pandas
python-dotenv
openai
google-generativeai
pyqt5